+++ src/Argv/ArgvOptionsProvider.ts
@@ -14,26 +14,52 @@ export class ArgvOptionsProvider <T> implements OptionsProvider<T> {
     const lookup = this._lookup()
 
     let options: Partial<T> = {} as any
+    let positionalIndex = -1
+
+    const addOption = (option: keyof T) =>
+      options[option] = this.parser.option(stream, option)
 
     while (!stream.isEmpty) {
       const flag = stream.next()
 
-      if (!(flag in lookup)) {
+      if (flag in lookup) {
+        addOption(lookup[flag])
+        continue
+      }
+
+      if (this.parser.positions == null) {
         throw new Error(`Unrecognized flag: ${flag}`)
       }
 
-      const prop = lookup[flag]
+      stream.previous()
+      positionalIndex++
 
-      options[prop] = this.parser.option(stream, prop)
+      if (this._acceptsPositional(positionalIndex)) {
+        addOption(this.parser.positions[positionalIndex])
+        continue
+      }
+
+      throw new Error(`Too many arguments`)
     }
 
     return options
   }
 
+  private _acceptsPositional (index: number): boolean {
+    return this.parser.positions != null &&
+      index < this.parser.positions.length
+  }
+
   private _lookup (): { [flag: string]: keyof T } {
-    return Object.keys(this.parser.flags)
+    const { flags: names } = this.parser
+
+    if (names == null) {
+      return {}
+    }
+
+    return Object.keys(names)
       .reduce((lookup, name: keyof T) => {
-        const flags = this.parser.flags[name]
+        const flags = names[name]
         const newLookups = flags
           .reduce((l: any, flag: string) => ({
             ...l,