+++ src/Argv/ArgvParser.ts
@@ -2,6 +2,7 @@ import {Flags} from './Flags'
 import {ArgvStream} from './ArgvStream'
 
 export interface ArgvParser <T> {
-  readonly flags: Flags<T>
-  option (stream: ArgvStream, flag: keyof T): T[typeof flag]
+  readonly flags?: Flags<T>
+  readonly positions?: (keyof T)[]
+  option (stream: ArgvStream, option: keyof T): T[typeof option]
 }